refactor: enum constant패키지로 분리
